requirements.txt
requests
transformers
torch
streamlit
faiss-cpu
sentence-transformers
rank-bm25
numpy
tqdm
pandas
matplotlib
